Fix framebuffer characters 128..255

+++ src/kernlib/display.cpp
@@ -87,7 +87,7 @@ class FramebufferDisplay: public Display {
     } else if (c == '\t') {
       offset += 8 - ((offset % charWidth()) % 8);
     } else {
-      const uint8_t *fontsym = fb_font_8x16 + (size_t(c) * 16);
+      const uint8_t *fontsym = fb_font_8x16 + (size_t(uint8_t(c)) * 16);
       size_t offsetX = offset % charWidth();
       size_t offsetY = offset / charWidth();
       size_t pb = pixelBytes();
